Outline for React routing
