+++ static/index.html
@@ -7,9 +7,7 @@
     <title>Webcandy</title>
   </head>
   <body>
-    <h1 id="title">Webcandy</h1>
     <div id="content"></div>
-
     <script src="dist/bundle.js" type="text/javascript"></script>
   </body>
 </html>
