+++ static/js/components/NotFound.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export default class NotFound extends React.Component {
+    // TODO: Why is CSS being imported?
+    render() {
+        return <h1>404: Not found</h1>;
+    }
+}
