+++ static/js/index.jsx
@@ -1,7 +1,25 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
+import NotFound from './components/NotFound';
 
-import '../css/index.css';
+class Hello extends React.Component {
+    render() {
+        return <p>Hello routing</p>
+    }
+}
 
-ReactDOM.render(<App />, document.getElementById("content"));
+const router = (
+    <Router>
+        <div>
+            <Switch>
+                <Route exact path="/" component={App}/>
+                <Route path="/hello" component={Hello}/>
+                <Route component={NotFound}/>
+            </Switch>
+        </div>
+    </Router>
+);
+
+ReactDOM.render(router, document.getElementById("content"));
