+++ static/package.json
@@ -8,7 +8,9 @@
     "react": "^16.8.3",
     "react-bootstrap": "^1.0.0-beta.5",
     "react-color": "^2.17.0",
-    "react-dom": "^16.8.3"
+    "react-dom": "^16.8.3",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.3.4",
